pyproject.toml
[tool.poetry]
name = "jpeg-pyopengl-filter-prototype"
version = "0.1.0"
description = ""
authors = ["Yusuke Kawamoto <novogrammer@gmail.com>"]
readme = "README.md"
packages = [{include = "jpeg_pyopengl_filter_prototype"}]

[tool.poetry.dependencies]
python = "^3.11"
pyopengl = "^3.1.7"
pillow = "^10.0.0"
pygame = "^2.5.1"
numpy = "^1.25.2"
python-dotenv = "^1.0.0"
opencv-python = "^4.8.0.76"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
